requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
validators
pytest
httpx
redis
psycopg2-binary
alembic
pytest-mock
fakeredis